requirements.txt
gradio>=4.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.1
pandas>=1.3.0
PyPDF2>=2.0.0
plotly>=5.3.0
python-docx>=0.8.11
psycopg2-binary>=2.9.1
logging>=0.5.1.2 